autoforcad/ansible/services/jerknotes/mail/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-Login==0.6.2
psycopg2-binary==2.9.7
Werkzeug<=2.3.7